requirements.txt
apollo_modules==7.0.2
shapely
loguru
docker
websocket-client
numpy
rtree
matplotlib
pygame
eventlet
tqdm
omegaconf
python-socketio==5.11.4
python-engineio==4.10.1
flask==3.0.3
flask-socketio==5.4.1
hydra-core==1.3.2
networkx
scipy==1.13.1